feat : 시설예약 기능구현

+++ src/main/java/com/mycompany/project/reservation/query/dto/FacilityDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 @AllArgsConstructor
 @ToString
 public class FacilityDTO {
-  private int facilityId;
+  private Long facilityId;
   private String name;
   private String status;
 }
